# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Printly document printing platform.",
        "title": "Printly API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Kimba SABI N'GOYE",
            "email": "kimbasabingoye@printly.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/centers/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all print centers awaiting approval. Requires admin role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get all pending print centers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.PrintCenter"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch pending centers",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/centers/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a print center. Requires admin role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a print center",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Print Center ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Print center deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Print center not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete print center",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/centers/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Approves, rejects, or suspends a print center. Requires admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update a print center's status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Print Center ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePrintCenterStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Print center not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update status",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/centers": {
            "get": {
                "description": "Retrieves a list of all approved print centers.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Print Centers"
                ],
                "summary": "Get all public print centers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.PrintCenter"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch print centers",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers a new print center, which will be pending approval. Requires authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Print Centers"
                ],
                "summary": "Register a new print center",
                "parameters": [
                    {
                        "description": "Print Center to create",
                        "name": "center",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePrintCenterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.PrintCenter"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to register print center",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/centers/{id}": {
            "get": {
                "description": "Retrieves a single print center by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Print Centers"
                ],
                "summary": "Get a print center by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Print Center ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.PrintCenter"
                        }
                    },
                    "404": {
                        "description": "Print center not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch print center",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a print center's information. Requires owner or admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Print Centers"
                ],
                "summary": "Update a print center's profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Print Center ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Print Center data to update",
                        "name": "center",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePrintCenterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Print center updated",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Print center not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update print center",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreatePrintCenterRequest": {
            "type": "object",
            "required": [
                "email",
                "location",
                "name",
                "phone_number",
                "working_hours"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/entity.Location"
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                },
                "phone_number": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Service"
                    }
                },
                "working_hours": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/entity.WorkingHour"
                    }
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "A description of the error"
                }
            }
        },
        "dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
...
        "dto.UpdatePrintCenterRequest": {
            "type": "object",
            "properties": {
                "location": {
                    "$ref": "#/definitions/entity.Location"
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                },
                "phone_number": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/entity.Service"
                    }
                },
                "working_hours": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/entity.WorkingHour"
                    }
                }
            }
        },
        "dto.UpdatePrintCenterStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "enum": [
                        "pending",
                        "approved",
                        "rejected",
                        "suspended"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.PrintCenterStatus"
                        }
                    ]
                }
            }
        },
...
        "entity.GeoPoint": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                }
            }
        },
        "entity.Location": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "geo_point": {
                    "$ref": "#/definitions/entity.GeoPoint"
                },
                "number": {
                    "description": "Postal address",
                    "type": "integer"
                },
                "street": {
                    "type": "string"
                },
                "type": {
                    "description": "e.g., \"Rue\", \"Avenue\"",
                    "type": "string"
                }
            }
        },
...
        "entity.PrintCenter": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "description": "gorm.Model is replaced to be explicit for swagger",
                    "type": "integer"
                },
                "location": {
                    "$ref": "#/definitions/entity.Location"
                },
                "location_id": {
                    "description": "Internal FK",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner_uid": {
                    "description": "Firebase UID of the manager",
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Service"
                    }
                },
                "status": {
                    "$ref": "#/definitions/entity.PrintCenterStatus"
                },
                "working_hours": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.WorkingHour"
                    }
                }
            }
        },
        "entity.PrintCenterStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected",
                "suspended"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusApproved",
                "StatusRejected",
                "StatusSuspended"
            ]
        },
...
        "entity.Service": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "description": "e.g., \"printing\"",
                    "type": "string"
                },
                "paper_size": {
                    "description": "e.g., \"A4\"",
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "entity.WorkingHour": {
            "type": "object",
            "properties": {
                "day": {
                    "description": "e.g., \"Monday\"",
                    "type": "string"
                },
                "end": {
                    "description": "Format: \"08:00\"",
                    "type": "string"
                },
                "start": {
                    "description": "Format: \"08:00\"",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and a JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}